Tests/LibWeb/Text/expected/hit_testing/content-editable-empty-block-descendants.txt

<P id="first">
<P id="first">
<P id="second">
hello
<#text>
